WINTERMUTE: Fix autodetection when adding a game.
Fix a bug introduced in a25b462889541ce5453955ca748985e4ab5079f8 that broke autodetection when adding a game.

1. Reinstate fileMan->registerPackages(fslist); in Wintermute::getGameInfo, as it *is* needed when adding a game.
2. Tackle the issue the previous commit was meant to fix one layer above, in advancedDetector.cpp.
An empty fslist is passed to fallbackDetect when starting a game that's already been added so that we don't pass the same files twice.
The piece of code in question is not called when adding a game.